OSIS-4203 moved LU detail view tests to own test class
	modified:   base/templates/learning_unit/search/base.html
	new file:   base/tests/views/learning_units/test_detail.py
	modified:   base/tests/views/test_learning_unit.py
