+++ base/templates/learning_unit/search/base.html
@@ -206,6 +206,11 @@
     <script>
         $(document).ready(function()
         {
+            {% if page_obj.start_index %}
+                const startingIndex = {{ page_obj.start_index }} - 1;
+            {% else %}
+                const startingIndex = 0;
+            {% endif %}
             columnDefs = [
                 {
                     "name": "academic_year",
@@ -222,7 +227,17 @@
                     "targets": 1,
                     "data": "acronym",
                     "render": function ( data, type, row, meta ) {
-                        return '<a href="'+ row['osis_url'] +'">'+ data + '</a>';
+                        const search_query_params = new URLSearchParams(window.location.search);
+                        search_query_params.set("index", startingIndex + meta["row"]);
+
+                        const params = new URLSearchParams();
+                        params.set("search_query", encodeURIComponent(search_query_params.toString().replace("?", "")));
+
+                        const url = `${row['osis_url']}?${params.toString()}`;
+                        const anchor = document.createElement("a");
+                        anchor.setAttribute("href", url);
+                        anchor.textContent = data;
+                        return anchor.outerHTML;
                     }
                 },
                 { "name": "title", "targets": 2, "data": "title"},
