+++ base/tests/views/test_learning_unit.py
@@ -597,193 +597,6 @@ class LearningUnitViewTestCase(TestCase):
         self.assertTemplateUsed(response, 'learning_unit/search/base.html')
         self.assertEqual(response.context['learning_units_count'], 1)
 
-    def test_learning_unit_read(self):
-        learning_container_year = LearningContainerYearFactory(academic_year=self.current_academic_year)
-        learning_unit_year = LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                                     learning_container_year=learning_container_year,
-                                                     subtype=learning_unit_year_subtypes.FULL)
-
-        header = {'HTTP_REFERER': SEARCH_URL_PART}
-        response = self.client.get(reverse('learning_unit', args=[learning_unit_year.pk]), **header)
-
-        self.assertTemplateUsed(response, 'learning_unit/identification.html')
-        self.assertEqual(response.context['learning_unit_year'], learning_unit_year)
-        self.assertEqual(self.client.session['search_url'], SEARCH_URL_PART)
-
-    def test_learning_unit_read_versions(self):
-        learning_unit_year = LearningUnitYearFullFactory(
-            academic_year=self.current_academic_year,
-        )
-        LearningComponentYearFactory(learning_unit_year=learning_unit_year)
-
-        response = self.client.get(reverse('learning_unit', args=[learning_unit_year.pk]))
-        self.assertEqual(len(response.context['versions']), 0)
-
-        with reversion.create_revision():
-            learning_unit_year.learning_container_year.save()
-
-        response = self.client.get(reverse('learning_unit', args=[learning_unit_year.pk]))
-        self.assertEqual(len(response.context['versions']), 1)
-
-        with reversion.create_revision():
-            learning_unit_year.learningcomponentyear_set.first().save()
-
-        response = self.client.get(reverse('learning_unit', args=[learning_unit_year.pk]))
-        self.assertEqual(len(response.context['versions']), 2)
-
-    def test_external_learning_unit_read(self):
-        external_learning_unit_year = ExternalLearningUnitYearFactory(
-            learning_unit_year__subtype=learning_unit_year_subtypes.FULL,
-        )
-        learning_unit_year = external_learning_unit_year.learning_unit_year
-
-        response = self.client.get(reverse('learning_unit', args=[learning_unit_year.pk]))
-
-        self.assertTemplateUsed(response, 'learning_unit/identification.html')
-        self.assertEqual(response.context['learning_unit_year'], learning_unit_year)
-
-    def test_external_learning_unit_read_permission_denied(self):
-        learning_container_year = LearningContainerYearFactory(academic_year=self.current_academic_year)
-        learning_unit_year = LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                                     learning_container_year=learning_container_year,
-                                                     subtype=learning_unit_year_subtypes.FULL)
-        external_learning_unit_year = ExternalLearningUnitYearFactory(learning_unit_year=learning_unit_year)
-        learning_unit_year = external_learning_unit_year.learning_unit_year
-
-        a_user_without_perms = PersonFactory().user
-        client = Client()
-        client.force_login(a_user_without_perms)
-
-        response = client.get(reverse("learning_unit", args=[learning_unit_year.id]))
-        self.assertEqual(response.status_code, HttpResponseForbidden.status_code)
-        self.assertTemplateUsed(response, "access_denied.html")
-
-        a_user_without_perms.user_permissions.add(
-            Permission.objects.get(codename='can_access_externallearningunityear'))
-
-        response = client.get(reverse("learning_unit", args=[learning_unit_year.id]))
-        self.assertEqual(response.status_code, 200)
-
-    def test_warnings_learning_unit_read(self):
-        learning_container_year = LearningContainerYearFactory(
-            academic_year=self.current_academic_year,
-            container_type=learning_container_year_types.INTERNSHIP,
-            requirement_entity=EntityVersionFactory().entity,
-            allocation_entity=EntityVersionFactory().entity,
-            additional_entity_1=EntityVersionFactory().entity,
-            additional_entity_2=EntityVersionFactory().entity,
-        )
-        LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                learning_container_year=learning_container_year,
-                                internship_subtype=internship_subtypes.TEACHING_INTERNSHIP,
-                                subtype=learning_unit_year_subtypes.FULL,
-                                periodicity=learning_unit_year_periodicity.BIENNIAL_ODD,
-                                status=False)
-        partim_without_internship = LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                                            learning_container_year=learning_container_year,
-                                                            internship_subtype=None,
-                                                            subtype=learning_unit_year_subtypes.PARTIM,
-                                                            periodicity=learning_unit_year_periodicity.ANNUAL,
-                                                            status=True)
-
-        response = self.client.get(reverse('learning_unit', args=[partim_without_internship.pk]))
-
-        self.assertTemplateUsed(response, 'learning_unit/identification.html')
-        self.assertEqual(len(response.context['warnings']), 3)
-
-    def test_learning_unit_with_faculty_manager_when_can_edit_end_date(self):
-        learning_container_year = LearningContainerYearFactory(
-            academic_year=self.current_academic_year,
-            container_type=learning_container_year_types.OTHER_COLLECTIVE,
-            requirement_entity=EntityFactory(),
-        )
-        learning_unit_year = LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                                     learning_container_year=learning_container_year,
-                                                     subtype=learning_unit_year_subtypes.FULL)
-        EntityVersionFactory(entity=learning_container_year.requirement_entity)
-
-        learning_unit_year.learning_unit.end_year = None
-        learning_unit_year.learning_unit.save()
-        ue_manager = UEFacultyManagerFactory(
-            'can_edit_learningunit',
-            'can_access_learningunit',
-            'can_edit_learningunit_date'
-        )
-        managers = [
-            FacultyManagerFactory('can_edit_learningunit', 'can_access_learningunit', 'can_edit_learningunit_date'),
-            ue_manager
-        ]
-
-        for manager in managers:
-            PersonEntityFactory(
-                entity=learning_container_year.requirement_entity,
-                person=manager
-            )
-            url = reverse("learning_unit", args=[learning_unit_year.id])
-            self.client.force_login(manager.user)
-
-            response = self.client.get(url)
-            self.assertEqual(response.context["can_edit_date"], True)
-
-    def test_learning_unit_of_type_partim_with_faculty_manager(self):
-        learning_container_year = LearningContainerYearFactory(
-            academic_year=self.current_academic_year,
-            container_type=learning_container_year_types.COURSE,
-            requirement_entity=EntityFactory(),
-        )
-        LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                learning_container_year=learning_container_year,
-                                subtype=learning_unit_year_subtypes.FULL)
-        learning_unit_year = LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                                     learning_container_year=learning_container_year,
-                                                     subtype=learning_unit_year_subtypes.PARTIM)
-        EntityVersionFactory(entity=learning_container_year.requirement_entity)
-        learning_unit_year.learning_unit.end_year = None
-        learning_unit_year.learning_unit.save()
-        ue_manager = create_person_with_permission_and_group(UE_FACULTY_MANAGER_GROUP, 'can_edit_learningunit')
-        ue_manager.user.user_permissions.add(Permission.objects.get(codename='can_access_learningunit'))
-        managers = [
-            create_person_with_permission_and_group(FACULTY_MANAGER_GROUP, 'can_edit_learningunit'),
-            ue_manager
-        ]
-        for manager in managers:
-            manager.user.user_permissions.add(Permission.objects.get(codename='can_edit_learningunit_date'))
-            manager.user.user_permissions.add(Permission.objects.get(codename='can_access_learningunit'))
-            PersonEntityFactory(entity=learning_container_year.requirement_entity, person=manager)
-            url = reverse("learning_unit", args=[learning_unit_year.id])
-            self.client.force_login(manager.user)
-
-            response = self.client.get(url)
-            self.assertEqual(response.context["can_edit_date"], True)
-
-    def test_learning_unit_with_faculty_manager_when_cannot_edit_end_date(self):
-        learning_container_year = LearningContainerYearFactory(
-            academic_year=self.current_academic_year,
-            container_type=learning_container_year_types.COURSE,
-            requirement_entity=EntityFactory(),
-        )
-        learning_unit_year = LearningUnitYearFactory(academic_year=self.current_academic_year,
-                                                     learning_container_year=learning_container_year,
-                                                     subtype=learning_unit_year_subtypes.FULL)
-        EntityVersionFactory(entity=learning_container_year.requirement_entity)
-        learning_unit_year.learning_unit.end_year = None
-        learning_unit_year.learning_unit.save()
-        managers = [
-            FacultyManagerFactory('can_access_learningunit'),
-        ]
-        for manager in managers:
-            PersonEntityFactory(entity=learning_container_year.requirement_entity, person=manager)
-            url = reverse("learning_unit", args=[learning_unit_year.id])
-            self.client.force_login(manager.user)
-
-            response = self.client.get(url)
-            self.assertEqual(response.context["can_edit_date"], False)
-
-    def test_get_components_no_learning_container_yr(self):
-        luy_without_components = LearningUnitYearFactory(academic_year=self.current_academic_year)
-        components_dict = learning_unit_business.get_same_container_year_components(luy_without_components)
-        self.assertEqual(len(components_dict.get('components')), 0)
-
     def test_get_components_with_classes(self):
         l_container = LearningContainerFactory()
         l_container_year = LearningContainerYearFactory(academic_year=self.current_academic_year,
@@ -797,40 +610,6 @@ class LearningUnitViewTestCase(TestCase):
         components_dict = learning_unit_business.get_same_container_year_components(learning_unit_year)
         self.assertEqual(len(components_dict.get('components')), 1)
         self.assertEqual(len(components_dict.get('components')[0]['learning_component_year'].classes), 2)
-
-    def test_get_partims_identification_tabs(self):
-        learning_unit_container_year = LearningContainerYearFactory(
-            academic_year=self.current_academic_year
-        )
-        learning_unit_year = LearningUnitYearFactory(
-            acronym="LCHIM1210",
-            learning_container_year=learning_unit_container_year,
-            subtype=learning_unit_year_subtypes.FULL,
-            academic_year=self.current_academic_year
-        )
-        LearningUnitYearFactory(
-            acronym="LCHIM1210A",
-            learning_container_year=learning_unit_container_year,
-            subtype=learning_unit_year_subtypes.PARTIM,
-            academic_year=self.current_academic_year
-        )
-        LearningUnitYearFactory(
-            acronym="LCHIM1210B",
-            learning_container_year=learning_unit_container_year,
-            subtype=learning_unit_year_subtypes.PARTIM,
-            academic_year=self.current_academic_year
-        )
-        LearningUnitYearFactory(
-            acronym="LCHIM1210F",
-            learning_container_year=learning_unit_container_year,
-            subtype=learning_unit_year_subtypes.PARTIM,
-            academic_year=self.current_academic_year
-        )
-
-        response = self.client.get(reverse('learning_unit', args=[learning_unit_year.pk]))
-
-        self.assertTemplateUsed(response, 'learning_unit/identification.html')
-        self.assertEqual(len(response.context['learning_container_year_partims']), 3)
 
     def test_learning_unit_formation(self):
         learning_unit_year = LearningUnitYearFactory(academic_year=self.current_academic_year,
